+++ libs/vscode/nx-cloud-fix-webview/src/nx-cloud-fix-component.ts
@@ -689,7 +689,7 @@ export class NxCloudFixComponent extends LitElement {
       }
 
       .codicon-modifier-spin {
-        animation: spin 1s linear infinite;
+        animation: spin 2s linear infinite;
       }
 
       .task-id {
